backend/pyproject.toml
[project]
name = "hanswer-backend"
version = "0.1.0"
description = "HAnswer — learning companion backend (FastAPI + Gemini + Milvus + PostgreSQL)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pymilvus>=2.4",
    "google-genai>=0.3",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.1",
    "tenacity>=8.3",
    "tomli>=2.0; python_version<'3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "mypy>=1.10",
]
retrieval = [
    # BAAI/bge-m3 dense + sparse embeddings (M5, §3.4).
    # Only needed if retrieval.embedder or retrieval.sparse_encoder is "bge-m3".
    "FlagEmbedding>=1.2",
]

[project.scripts]
hanswer-prompt = "app.prompts.cli:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
